chore: move OL components to ol directory
